Extract onClickListener from factory
