+++ app/src/main/kotlin/com/saltedge/authenticator/features/settings/mvvm/about/AboutItemViewHolder.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of the Salt Edge Authenticator distribution
  * (https://github.com/saltedge/sca-authenticator-android).
- * Copyright (c) 2019 Salt Edge Inc.
+ * Copyright (c) 2020 Salt Edge Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -35,12 +35,9 @@ class AboutItemViewHolder(
     }
 
     fun bind(item: SettingsItemViewModel, listener: OnItemClickListener?) {
-        aboutBinding?.about = AboutListItemViewModel(item)
+        aboutBinding?.aboutListItem = AboutListItem(item)
         aboutBinding?.onItemClickListener = listener
         aboutBinding?.executePendingBindings()
     }
 }
 
-interface OnItemClickListener {
-    fun onItemClick(titleName: Int)
-}
