+++ app/src/main/kotlin/com/saltedge/authenticator/features/settings/mvvm/about/AboutViewModel.kt
@@ -1,3 +1,23 @@
+/*
+ * This file is part of the Salt Edge Authenticator distribution
+ * (https://github.com/saltedge/sca-authenticator-android).
+ * Copyright (c) 2020 Salt Edge Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3 or later.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * For the additional permissions granted for Salt Edge Authenticator
+ * under Section 7 of the GNU General Public License see THIRD_PARTY_NOTICES.md
+ */
 package com.saltedge.authenticator.features.settings.mvvm.about
 
 import android.content.Context
@@ -6,7 +26,7 @@ import com.saltedge.authenticator.R
 import com.saltedge.authenticator.features.settings.common.SettingsItemViewModel
 import com.saltedge.authenticator.tool.AppTools
 
-class AboutViewModel(val appContext: Context, val onItemClickListener: OnItemClickListener): ViewModel() {
+class AboutViewModel(val appContext: Context): ViewModel() {
 
     fun getListItems(): List<SettingsItemViewModel> {
         return listOf(
